pintrail/portal/pyproject.toml
[project]
name = "pt2"
version = "0.1.0"
description = "Pintrail portal v2 - Python/FastAPI/SQLModel/HTMX"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0,<0.116.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.30.0",
    "greenlet>=3.1.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.34.0",
    "pillow>=11.0.0",
    "arq>=0.26.0",
    "aiofiles>=24.0.0",
]

[project.scripts]
server = "app.main:run"
worker = "worker:run"

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
